Create insertBand.php
php to insert a new band into mysql database - BandID is null to auto increment the id number, date was added in hard code, so I didn't have to worry about figuring out the date field yet.  This was just for an initial run to make sure that data could be inserted.